+++ semantic-conventions/src/tests/semconv/templating/test_code.py
@@ -1,7 +1,5 @@
 import io
 
-import pytest
-
 from opentelemetry.semconv.model.semantic_convention import SemanticConventionSet
 from opentelemetry.semconv.templating.code import CodeRenderer
 
@@ -12,7 +10,7 @@ def test_codegen_units(test_file_path, read_test_file):
     semconv.finish()
 
     template_path = test_file_path("jinja", "metrics", "units_template")
-    renderer = CodeRenderer({}, {})
+    renderer = CodeRenderer({}, False)
 
     output = io.StringIO()
     renderer.render(semconv, template_path, output, None)
@@ -23,21 +21,14 @@ def test_codegen_units(test_file_path, read_test_file):
     assert result == expected
 
 
-@pytest.mark.parametrize(
-    "trim_blocks, lstrip_blocks", [("True", "True"), ("true", "TRUE"), ("TRUE", "tRuE")]
-)
-def test_strip_blocks_enabled(
-    test_file_path, read_test_file, trim_blocks, lstrip_blocks
-):
-    """Tests that Jinja config options are passed to the Jinja environment"""
+def test_strip_blocks_enabled(test_file_path, read_test_file):
+    """Tests that the Jinja whitespace control params are fed to the Jinja environment"""
     semconv = SemanticConventionSet(debug=False)
     semconv.parse(test_file_path("yaml", "metrics", "units.yaml"))
     semconv.finish()
 
     template_path = test_file_path("jinja", "metrics", "units_template")
-    renderer = CodeRenderer(
-        {}, {"trim_blocks": trim_blocks, "lstrip_blocks": lstrip_blocks}
-    )
+    renderer = CodeRenderer({}, True)
 
     output = io.StringIO()
     renderer.render(semconv, template_path, output, None)